+++ upyog/utils/__init__.py
@@ -2,3 +2,4 @@ from .utils import *
 from .boxes import *
 from .download import *
 from .zip_utils import *
+from .prettify_df import *